Authorization & Error Handling
Everything's done, time for amazon web services.

+++ backend/routes/posts.js
@@ -36,18 +36,24 @@ router.post(
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath:  url + "/images/" + req.file.filename
+    imagePath:  url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
-  console.log(post)
-  post.save().then(createdPost => {
-    console.log(createdPost)
+  post
+  .save()
+  .then(createdPost => {
     res.status(201).json({
       message: 'Post added successfully',
       post: {
         ...createdPost,
         id: createdPost._id,
       }
-      });
+    });
+  })
+    .catch(error => {
+      res.status(500).json({
+       message: "Creating a post failed(unknown error)"
+     });
   });
 });
 
@@ -83,13 +89,21 @@ multer({storage: storage}).single("image"), (req, res, next) => {
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   })
-  console.log(post);
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    console.log(result);
-    res.status(200).json({message: 'Update succeeded!'});
-  });
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+    if (result.nModified > 0){
+      res.status(200).json({message: 'Update succeeded!'})}
+    else{
+      res.status(401).json({ message: "Update post failed(you are unauthorized)" })
+    }
+    })
+    .catch(error => {
+      res.status(500).json({
+        message: "Update post failed(database may be down)"
+      });
+    })
 });
 
 router.get("/:id", (req, res, next) =>{
@@ -99,15 +113,33 @@ router.get("/:id", (req, res, next) =>{
     } else {
       res.status(404).json({message: "WHERE POST AAAAAAA"})
     }
+  })
+  .catch(error => {
+      if (post) {
+        res.status(200).json(post);
+      } else {
+        res.status(404).json({ message: "Post not found(ID may be wrong) "});
+      }
+    }).catch(error => {
+      res.status(500).json({
+        message: "Fetching post failed(server may be down)"
+      });
+    });
   });
 
-})
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Post kicked the bucket" });
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+    if (result.n > 0){
+      res.status(200).json({message: 'Deletion Successful'})}
+    else{
+      res.status(401).json({ message: "Not authorized!" });
+    }
+  }
+).catch(error => {
+  res.status(500).json({
+   message: "Deleting post failed(server may be down?)"
+    });
   });
 });
-
 module.exports = router;
